chore: Update generator.js to handle different database types
The generator.js file has been updated to handle different database types based on user input. If the selected database is not "mongoose", the user is prompted to choose a database type (mysql, postgres, sqlite, mssql), and the handleDatabaseCreation function is called with the selected database type and project name. This change improves the flexibility of the project creation process by allowing users to choose their preferred database type.