feat: Add Mongoose integration to handle different database types
This commit adds the ability to handle different database types in the generator.js file. If the selected database is "mongoose", the mongooseIntegration function is called with the project name. This allows users to choose Mongoose as their preferred database type during project creation. The addition of Mongoose integration improves the flexibility of the project creation process.