feat: Add Drizzle integration for MySQL and PostgreSQL databases
This commit adds the ability to integrate Drizzle ORM with MySQL and PostgreSQL databases in the generator.js file. If the selected database is "mysql", the drizzleSqlConfig function is called with the project name. If the selected database is "postgres", the drizzlePostgreConfig function is called. This allows users to choose Drizzle as their preferred database type during project creation. The addition of Drizzle integration improves the flexibility of the project creation process.