feat: Add command to add a database to the project
This commit adds a new command to the CLI tool that allows users to add a database to their project. The command is named "add:database" and when executed, it logs a message indicating that a database is being added to the project. This feature enhances the functionality of the CLI tool by providing users with the ability to easily incorporate a database into their projects.